Shell: Put background jobs into new process groups too
Fixes #4345.
This was done in 8ea641f in the name of "fixing event loop processing in
subshells", but I do not see how a new PGID is supposed to affect the event
loop.
This seems to have been done by mistake, let's see if any tests fail
because of this.