BXVGADevice+MBVGADevice: Correctly check page-aligned mmaps
In 77ba5fa1419a743490b344192b73b9db5babe2f3, mmap was changed so that
the size of the region is aligned before it was passed to the device
driver. The previous logic would assert when the framebuffer size was
not a multiple of the page size. I've also taken the liberty of
returning an error on mmap failure rather than asserting.